Match multiple stage-related modules

+++ src/game/time_attack/results.c
@@ -50,7 +50,7 @@ u32 CreateTimeAttackResultsCutScene(u32 finishTime)
     u8 i;
     u8 isBossLevel;
     u8 level;
-    gLoadedSaveGame->score += gRingCount;
+    gLoadedSaveGame->score += (s16)gRingCount;
 
     t = TaskCreate(sub_8089AEC, sizeof(struct TimeAttackResultsCutScene), 0xC100, 0,
                    sub_8089BB0);
